necessary imports for fancy movement + clean up
Buttons are currently locked to top left for now.
The buttons now also have a shake-y effect applied that will be easier and neater than the former gradient drop shadow that used to be applied.
Some ancient code clean up was also done, so it should look a good amount tidier.